RFCT Move functionality to package level
This would enable this to be used as a Python package
